apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.1.1
boto3==1.29.7
cloudflare==2.11.7
openai==1.3.5
replicate==0.22.0
redis==5.0.1
celery==5.3.4
stripe==7.5.0
supabase==2.0.3
ffmpeg-python==0.2.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
moviepy==1.0.3
pytest==7.4.3
black==23.11.0
PyJWT==2.8.0
bcrypt==4.1.2
pydantic[email]==2.5.0
aiohttp==3.9.1
aiosqlite==0.19.0
langdetect==1.0.9
elevenlabs==0.2.26
deep-translator==1.11.4